Construct StoreFactory in CLI
This change creates the store factory in CLI instead of using
libsplinter's create_store_factory function.  This pushes the
construction of the factory closer to the decision point of which
implementation to use.

The change removes the use of a Memory connection URI, as it is not
possible to use that in the CLI.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>