Update admin routes to error if protocol version is too high
Before the protocol version check had a catch all that would
default to the current protocol version. While this is safe
because there is already a guard around the route that would
reject a protocol that is too high, this commit adds an
explicit check in case the guard fails.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>