Add level field to AppenderConfig structs
In earlier versions of AppenderConfigs and all the related structs there
wasn't an option to set log level filters. Now that the option is being
added there needs to be a way to configure that. This commit adds an
optional level field to AppenderConfigs and updates places that directly
construct those structs.

Signed-off-by: Caleb Hill <hill@bitwise.io>