Add logic for appender filter config
If there is a level configured on an AppenderConfig that level needs to
be reflected in the log4rs Appender that gets created. The code in this
commit checks if a level is set, converts it to the type needed by the
log4rs AppenderBuilder, and adds it to the builders lists of filters.

Level -> LevelFilter -> ThresholdFilter
                               |
                               v
                        AppenderBuilder

Signed-off-by: Caleb Hill <hill@bitwise.io>