Add receipt store upgrade module
Add a module which contains the methods to upgrade the receipt store by
locating any LMDB receipt store files and migrating the data to the new
sqlite or postgres backed receipt store. After migrating, '.old' is
added to the end of all LMDB receipt store file names so that data will
not be migrated multiple times if the upgrade command is run again.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>