Add default value logic to LoggerConfig
As part of the effort to make configuring logging easier more fields are
getting "defaults". The `level` field will get the root loggers level as
a default and the `appenders` field will get an empty vec as a default.
Since loggers are additive most loggers don't need their own lists of
appender targets as they will also log to the root logger.

Signed-off-by: Caleb Hill <hill@bitwise.io>