Fix update_proposal so it does not remove incorrect nodes
When updating a proposal, the command was removing on proposed
nodes with the same node id. This is incorrect because there maybe
several proposed nodes with different circuits. This commit
updates the command to remove the nodes by circuit id instead.

This was only seen when submitting multiple 3 party circuit
transactions and submitting only one vote for each proposal.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>