Create Database file with 640 permissions
There are two related changes in the commit.

1. Creating the file if it does not exist already
2. Setting the mode to 640 for rw-r----- file permisssions

If the file does not exist when the filename gets passed to the
SQLiteConnection constructor the Rust library that handles the foreign
function interface code to the sqlite3 C++ library hard codes 644
(rw-r--r--) as the file permissions. To avoid that the file just has to
exist beforehand with the permissions we want.

Signed-off-by: Caleb Hill <hill@bitwise.io>