Add a simple integration test that triggers a peer interconnect bug
If there is a message sent before the duplicate connection is replaced,
the admin message may be dropped and the peer id to connection id in
the peer interconnect may not be correctly updated.

This test is currently ignored until the bug can be fixed to the
test reliable passes.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>