Skip receipt store upgrade if 0 circuits
Add a check to receipt store upgrade that will skip the receipt store
upgrade process if there are no local circuits found. Previously, if
there were no local circuits found the receipt store upgrade would run
into an error when creating the the `LmdbReceiptStore` because it
attempted to get the first file name in the vector but couldn't
because the vec was empty.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>