Rename config value for scabbard storage
This change renames the clap value for `scabbard_storage` to be
`scabbard_state`.  This both matches the CLI flag, as well as matches
the intent of the configuration value: it only modifies scabbard's
storage of state specifically.

This change also renames the module, enum and error to an equivalent
naming scheme.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>