Check for duplicate connection when adding unreferenced peer
There is an unlikely chance that, a peer will request an
unreferenced peer and received an unreferenced peer from the
same node. If this has happened there is a chance that when
upgrading the outbound connection to a Peer, which
connection to create the peer for with.

Before this was causing one side to close the connection and
the other side to be unaware of the removal until it tried
to send a message.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>