Remove `lower_snake_case` property validation
Batch endpoint fails this validation, can not be excluded due to
limitations in the linter, and can not be changed for compatibility
reasons. This change disables the entire lint. Ideally we would submit a
fix to IBM's OpenAPI linter to allow exclusion in the future or come
back and fix the endpoint so we can enable the lint.

Signed-off-by: Lee Bradley <bradley@bitwise.io>