Fix calculation in `to_milli`
Fix the calculation in the `to_milli` function to divide the amount of
time in milliseconds by the number of batches to be submitted in that
interval, to get the amount of time to wait between submissions.
The previous calculation, multiplying the number of milliseconds by the
number of batches to be submitted, gave an incorrect value when any value
other than 1/s was given on the command line for `target-rate`.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>