Fix rate calculation in `start_..._workloads`
Update both the `start_command_workloads` and
`start_smallbank_workloads` functions to use `target_rate_min` as the
rate if the minimum and maximum target rates are equal.

Also update the calculations used to get the rate if a range is given.
With the fixed calculations in `to_milli`, the max target rate returned
by `to_milli` will be smaller than the minimum rate because the value
returned by `to_milli` is the amount of time to wait between
submissions, so the updated calculations generate a random value between
(max..=min) and then calculate the number of batches that should be
submitted per second with this generated time.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>