Create default splinter-service feature
This change adds the "splinter-service" feature and adds it to the list
of default features.  Essentially this is a wrapping a currently default
code into a default feature, such that other items (like the client
code) can be pulled in on their own, without all of the requirements of
the service.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>