Use CircuitTemplate::load_from_file to use the file_path directly
Every time this is used it is passed the entire file path. The file
path was found from the default and env variables previously.

Trying to find a file based on the full file path would cause the
file to not be found, and thus the show and argument CLI commands
would fail.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>