Update factory guards with no db feature enabled
This change updates the all of the guards such that the code correctly
compiles and lints when no database features are enabled.  This causes a
number of guards to be added, but the current structure of the factory
does not easily support isolating these items.

As this code will be significantly re-written in upcoming releases, the
feature noise here is acceptable.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>