Simplify ByteSize to a wrapper around a u64
ByteSize exists to make entering file sizes like '5.5G' for five and a
half gigabytes easier than counting out the right number of zeros.
Ultimately it gets converted to a u64 because thats what the filesystem
module uses for file sizes.
This commit simplifies the ByteSize struct to just be a wrapper around a
u64 with all the math being done around deserialization time.

Signed-off-by: Caleb Hill <hill@bitwise.io>