Add default for rolling_file size
The default is set at 100Mb. The default value is set in the
config::logger module as opposed to config::toml because config::toml is
only concerned with deserialization and not providing defaults.  While
config::logger deals with all the details of converting a deserialized
struct to one we can actually use.

Signed-off-by: Caleb Hill <hill@bitwise.io>