Move encoder default value out of toml module
The toml structs should only be concerned with deserialization of the
toml file. Including logic for setting field defaults is outside that
scope.

This commit moves the setting of the encoders default value out of the
config::toml module and into the config::logging module because that
module deals with the internal representation of log config and is a
more appropriate place to set default values.

Signed-off-by: Caleb Hill <hill@bitwise.io>