Replace default const with LogEncoder default impl
The value of `DEFAULT_LOGGING_PATTERN` is used a bunch of places to create
default appenders. This change replaces all those uses with
`LogEncoder::default()`.

The implementation of Appender encoder is hardcoded at the moment but
might not always be. Adding a type with a default impl makes switching
out the implementation easier and less prone to error.

Signed-off-by: Caleb Hill <hill@bitwise.io>