Ensure Committed BatchInfo sent to subscribers
This change updates the scabbard state to ensure that Committed
BatchInfo is sent to the subscribers.  Previously, there was a mix
between Valid and Invalid or Committed and Invalid.

Additionally, Committed statuses were not updated via the same mechanism
as other statuses, so those BatchInfo instances were never sent to the
subscribers.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>