Add From impl to LogEncoder to boxed Encoder trait
LogEncoder decouples the toml representation of log encodings
from the actual dynamic Encoder the logging system uses. For now there
is no functional change, a LogEncoder always maps to a log4rs
PatternEncoder. But it makes if flexible enough it could return other
Encoder trait objects in the future if the pattern/encoder field's
meaning changes.

Signed-off-by: Caleb Hill <hill@bitwise.io>