Update flexi_logger dependency to v0.20
This is intended as a non-functional change the track the latest
dependency.

Adds the "use_chrono_for_offset" feature as required to avoid warnings
from this version of flexi_logger; this is a flexi_logger problem and
can be safely removed when its no longer needed to avoid warnings from
flexi_logger.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>