Remove health service
This change removes the health service, both from splinterd and the
repository in general.

The health service, as it stands, is merely skeleton code that prior to
this commit was providing no value and slowing down the build times.

As the service API will be undergoing radical redesign in future
releases, the health service should be re-introduced after those
designs have been finalized.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>