Add StateMigrateAction for moving scabbard state
Add Action implementation for moving scabbard data from or to an LMDB
database. This allows for moving the existing LMDB scabbard into the
SQLite or PostgreSQL database that the rest of the splinter state is
stored. It also provides a way to move scabbard state out of this
database into individual LMDB databases as well.

Only the current commit hash is kept, the rest are pruned. After the
data has properly been moved, the tree is deleted from the input
database.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>