Add auth to `Network`
Add an auth field to the `Network` struct as well as a
`with_client_auth` function which can be used to set this field when
creating a new test network. In the `Network`
`add_nodes_with_defaults` function pass the auth value into the
NodeBuilder using the node builder's `with_client_auth` function.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>