Update gameroom to use CircuitTemplateManager
This replaces the use of CircuitCreateTemplate that was used
previously and updates gameroom to follow the same pattern as
the splinter CLI.

Previously, the CLI was broken and when fixed, broke gamerooms
ability to find its template. Having them use the same method
allows them both to work without conflicts.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>