Add more default loggers
This change adds `sawtooth`, `transact` and `cylinder` loggers to the set of default
loggers. This means by they will log everything to whatever appenders
are configured on the root logger. This means the easiest/coarsest dial
to adjust how many logs an appender actually writes out is the appender
level.

In the stock configuration this means the verbosity flag
controls how many logs are written for all of these loggers.

Signed-off-by: Caleb Hill <hill@bitwise.io>