Add ConnectionPool enum
This change adds a connection pool enum that may be used by stores to
provide either standard pool usage or write-exclusive usage.

It provides two functions: execute_write and execute_read. Both
functions provide a connection to a closure.  If the ConnectionPool is
WriteExclusive, it will use a RwLock for each respeciive methods such
that it enforces the single writer multiple reader guarantee.

It does require several feature guard changes to make the store module
available under the write guard conditions. This includes a "store"
feature that is effectively grandfathered into stable.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>