Add write-exclusivity to Scabbard store
This change adds write-exclusivity to the Scabbard
DieselCommitHashStore. It follows a very similar pattern to libsplinter,
but without the generic error.

This uses its own version of the ConnectionPool, as the libsplinter
version isn't library public, and should allow the scabbard client to
diverge in its underlying behavior, while still accepting a
Arc<RwLock<...>> wrapped pool.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>