Add ConnectionPool enum
This change adds a connection pool enum that may be used by stores to
provide either standard pool usage or write-exclusive usage.

It provides two functions: execute_write and execute_read. Both
functions provide a connection to a closure.  If the ConnectionPool is
WriteExclusive, it will use a RwLock for each respective methods such
that it enforces the single writer multiple reader guarantee.

This struct is identical to the one in libsplinter.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>