Replace rust-crypto with sha2
This change replaces the rust-crypto library with the sha2 library.  The
latter is a smaller, single purpose library, as well as being compatible
as-is on wasm32.

The change does require the addition of converting the hash to a hex
string in a number of places.

In the case of the cli, rust-crypto was not used, so it has simply been
removed.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>