Ignore "database is locked" on pool conn creation
This change adds an error handler for the connection pool that will
ignore logging a "database is locked" error if it occurs while creating
the connection.

This error can occur due to the setting of PRAGMA statements in the
ConnectionCustomizer implementation.  The connection manager will re-try
creating the connection after a delay, so the error can safely be
ignored.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>