Change function signature of InvalidArgumentError::new()
This changes InvalidArgumentError::new() to take generic Into<String>
arguments instead of String directly. This prevents the caller from
needing to call to_string() or into() explicitly in several cases.

While this is mostly compatible with existing code, it is not compatible
when the calling code was using into() because the compiler can not
determine the intermediate type in that case.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>