Add sqlite migration for renamed role tables
sqlite does not always update the parent table references on table
renames. When the `roles` table was updated to be named `rbac_roles` the
foreign key constraint on the `role_permission' table was not updated
making it impossible to insert any new role:permission relations because
the foreign key constraint would fail.

This commit adds a database migration to update the foreign key
constraint on the `rbac_role_permissions` table to reference the
`rbac_roles` table instead of the `roles` table.

It also adds an undo sql migration that breaks it again.

Signed-off-by: Caleb Hill <hill@bitwise.io>