Modify splinter user list command
Modify the splinter user list command to be able to list just biome or
just oauth users. Previously, in the case that either the /biome/users or
/oauth/users endpoint wasn't available this CLI command would fail and
return an error saying the endpoint doesn't exist. This was an issue
when attempting to list users while only biome auth was configured
because the /oauth/users endpoint doesn't exist, causing the command
to return an error.

An error is still logged in case one or both of the auth types are
configured and the appropriate endpoint fails to return the list of
users.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>