Remove check excluding full filepath templates
Checking through the configured template directories for the passed in
filename is good when the user is using those directories. But when the
user passes in the full filepath to the template file it will
incorrectly reject the template path.

This commit changes to the code to always try to load the filepath as
given and then maps the errors to the previously returned errors.

Signed-off-by: Caleb Hill <hill@bitwise.io>