Remove all actix 1 related compression deps
This change removes all transient dependencies of actix-web, actix-http
and awc (actix web client) for compression libraries. It both ensures
that the transient dependency of brotli and the flate2 dependencies are
removed.

These older libraries' dependencies may cause conflicts with library
consumers that are using newer versions of the actix libraries.

As none of the current splinter clients use compressed responses, this
is safe to remove.  It may be provided in the future as the rest-api is
broken up into additional crates.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>