Fix error with missing root logger level from toml
The root logger should have a default level to be consistent with the
other loggers in the config file. So that you can ommit it in the config
file.

This commit adds a default level of `Warn` if the root logger
doesn't have a specified level. This matches the default config.

Signed-off-by: Caleb Hill <hill@bitwise.io>