Fix admin service protocol response sender for challenge auth
The incorrect peer token was being added to the protocol agreement
map, a trust id instead of challenge. Everything would continue to
act normally until a new proposals was added that used Trust. The
AdminService would think it was already connected to the node
via trust and sending the proposal would fail because the
connection does not exist.

This PR switches the sender of the protocol response to be the
correct format, admin::public_key::PUBLIC_KEY::public_key::PUBLIC_KEY,
when using challenge authorization which fixes this issue.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>