Recover from panics in thread pool jobs
This change introduces a Supervisor thread to the thread pool to watch
for messages about panicking workers.

Each worker has a monitor that will send a message if it detects that
the thread is panicking when it is dropped. If it is dropped, it can try
to send a message to the supervisor to create a new worker and replace
the old, failed one.

The supervisor is shutdown when the thread pool is shutdown.

Tests are added to validate proper behavior and panicking behavior.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>