Add resources mod
Resources are the code you think of running when you access an endpoint,
they operate on the database, convert data formats, and generally act
RESTfully. They are versioned, denoted here by `v1.rs` for version one
and `v2.rs` for version two. This code is entirely actix free and would
be preserved without modification between actix version upgrades.

All of the resources have public `Arguments` and `Response` structures
that correspond to a bundled collection of the needed arguments, and the
response structure that is whatever data was requested.

These structures are not shared between versions even if they contain
the same set of internals. This prevents messy internal dependencies
between versions. They are usable outside of the resouces module.

Signed-off-by: Caleb Hill <hill@bitwise.io>