Correct serialization so message is never dropped
Previously, if a batch status response was "Unknown" or "Pending", the
message field would not be returned (unlike "Invalid", "Valid" and
"Committed"). The OpenAPI spec states that message is not an optional
return value. This change modifies the serde serialization so that both
the "Unknown" and "Pending" enum variants are serialized as empty
sequences, and thus message is never empty.

This change also adds tests to validate the resulting JSON output. To
ensure that the string ouput for tests is deterministic, the
preserve_order feature is enabled in development builds.

Signed-off-by: Lee Bradley <bradley@bitwise.io>