Use ServiceDispatcher in CircuitDirectMessageHandler
This change adds the ServiceDispatcher to the
CircuitDirectMessageHandler, allowing for the use of
MessageHandler-based service implementations to handle the contents of a
circuit direct message payload.

This change updates the tests both to include tests for dispatching via
the service dispatcher as well as updating the circuit/service ID's to
match the format in the FullyQualifiedServiceId struct.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>