Fix needless_late_init lint
This change fixes a lint introduced with the Rust 1.59 release where a
variable is unnecessarily initialized late. That is, it can be
initialized via an if expression.

Additionally, an option was unwrapped (via an `if let Some...`
statement) and then re-assigned back to the variable.  This can be
replaced with using a reference to the option and not re-assigning at
all.

See
https://rust-lang.github.io/rust-clippy/master/index.html#needless_late_init
for more details.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>