Implement LifecycleExecutor
The LifecycleExecutor will check the database for pending
services that need to be prepared, finalized, retired or purged.

It will check the database on a configured interval or can be
woken up by an alarm. When it finds a pending service, it will
get the commands from the Lifecycle implementation for the service
type and add its own command to update the Lifecycle store. It
then passes all commands to an executor.

The commands should all be very quick so the LifecycleExecutor
only has one thread.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>