Rename and move Service to instance::ServiceInstance
This is the first in a series of commits to move the existing APIs for
services into a new modules to make room for the new threadless design.

"Instance" in this context represents the fact that implementations of
the trait only support a single instance of a service (one service id
per struct instance).

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>