Add paging module to common
Paging is a common operation and generally handled at the versioned
resource level and thus belongs in the common crate. This commits
contents are based on the libsplinter paging module. There are two main
changes to that libsplinter code.

1. The module structure is different, paging is versioned similarly to
   how grid handles pagination versions.

2. There is a builder instead of a single function to construct the
   Paging struct. This also lets the fields on the Paging struct be
   private so the programmer can't manually construct an invalid paging
   struct.

The Paging struct needs to implement Serialize and Deserialize as it is
generally part of the argument and return values for any paginated
resources. Actix and most other frameworks require those types implement
those serde traits.

Signed-off-by: Caleb Hill <hill@bitwise.io>