Add retry to sending messages in PeerInterconnect
This commit adds retry to sending messages to a connection by
updating the pending incoming messages loop to handle booth
incoming and outgoing messages.

Now if the PeerInterconnect receives a message that should be
sent to a Peer, but the message fails to send because the
connection has disconnected, the message will be sent to the
pending thread, where it will be retried up to 3 times, with
10 seconds in between each try.

The connection could go away because of race condition where,
duplicate connections are rectified but one side removes
the old connection before the other finishes authorization.
This change would allow for the message to be sent over the
new connection successfully.

Also fixes periodic test failures in integration test
test_2_party_circuit_duplicate_connection

Moves the pending loop into its own file.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>