Add Timer for new service API
The Timer is in charge of checking the configured
`TimerFilters` for pending work.

On start up the Timer takes a list of `TimerFilters`
and their associated `TimerHandlerFactory`. The factories
must return a handler that can handle messages of type `Vec<u8>`.

On wake up, the Timer will check each `TimerFilter` for pending work.
For each `FullyQualifiedServiceId` returned from the filter,
the associated `TimerHandlerFactory` will be used to get a
new `TimerHandler` that will run in a threadpool.

A `Pacemaker` is used to wake up the Timer on some interval.
An alarm is also provided to wake up the Timer prematurely.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>