Add ServiceDispatcher
This change adds the ServiceDispatcher, which dispatches byte message to
the appropriate message handler, constructed via the routable message
handler factory.

The message is handled within a task runner, a trait that has the
appropriate requirements to be backed a thread or thread pool.

The ServiceTypeResolver provides a trait to resolve a service type from
a given fully-qualified service ID.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>