Add internal Method to PermissionMap
With both actix_web_1 and actix_web_4 rest apis depending on the
permission map it can't incorporate bits of actix_web_1 in code used by
actix_web_4 code. This adds an internal Method representation with a
From impl for the Actix1Method enum.

This avoids making PermissionMap fully generic and in theory it
supports mixing and matching bits of both rest apis.

Signed-off-by: Caleb Hill <hill@bitwise.io>