Change PermissionMap visibility to public
As more and more of the rest api code moves to the rest_api crates they
will depend on PermissionMap. For the moment PermissionMap is
staying in libsplinter but it does need to be visible to those crates.
This commit changes the visibility of PermissionMap, its re-export in
the containing module, and adds the Method enum to that re-export.

Signed-off-by: Caleb Hill <hill@bitwise.io>