Set initial state root properly on start up of Scabbard
Previously, Scabbard did not set the initial state root
hash in the commit store on start. This caused the
`state migrate` command to fail if a scabbard circuit
was created but never used.

Scabbard will now properly set the state root on start.
If the state was not set previously, it will be set on
restart.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>