Add rest_api feature to libsplinter dep in common
This is the rest_api/common crate and it depends heavily on the rest_api
feature in libsplinter. This commit adds the rest_api feature as
part of the default feature set for the rest_api/common crate's
dependency declaration for libsplinter.

Once everything that eventually will end up in the common crate makes
its way there this change can be reversed.

Signed-off-by: Caleb Hill <hill@bitwise.io>