Add rest_api feature to actix_web_1
actix_web_1 needs to have a short term dependency on the rest_api
feature in libsplinter because otherwise you end up in a strange sort of
chicken egg problem with the existing actix web code in libsplinter.
Eventually this commit can be undone as more things get moved into
common and actix_web_1.

Signed-off-by: Caleb Hill <hill@bitwise.io>