Move CircuitStatus impl to non rest_api module
As the rest endpoint code moves out of libsplinter and into the
actix_web_1 crate any impl blocks for core libsplinter types contained
in those endpoint files need to be moved or rewritten. This commit moves
the From<String> implementation for CircuitStatus into the same module
as where CircuitStatus is defined.

Signed-off-by: Caleb Hill <hill@bitwise.io>