Update `list_consensus_actions` return type
Update the `list_consensus_actions` method to return a list of
`IdentifiedScabbardConsensusAction` so that the the associated
`action_id` is returned with the actions.

This change is necessary because a componenet executing the
actions would need the `action_id` in order to be able to
update the `executed_at` time for the actions in the database.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>