Update admin tests to not use insecure functions
These tests are moving out of libsplinter and into the actix_web_1
crate. The build_insecure function is defined in a test module which
can't and shouldn't be part of the libraries api. This commit updates
the tests being moved to use mocked IdentityProviders and
AuthorizationHandlers and the secure versions of build and run.

Signed-off-by: Caleb Hill <hill@bitwise.io>