Add consensus type to ScabbardService
A ScabbardService will be able to be run with different
types of consensus, configured from the circuit definition.

Currently, Scabbard V3 will support 2PC. This commit leaves
room for adding more supported consensus algorithms in the future.

Storing this in the scabbard service is required so the consensus
messages can be deserialized at runtime.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>