Remove service-network feature
The components under this feature are not compatible with
the new Service API, as such it will not be completed.

The current implementation has timing issues where a message
may be received from an unknown service because the service
connection manager has not finished handling notifications
from the connection manager about the services connection.
This has started causing test failures on the faster CI
machines.

This issues was also present in the PeerManager and fixed by
implementing retries.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>