Add ProposalStoreFactory trait to libsplinter
This commit primarily adds a ProposalStoreFactory trait that defines the behaviour
of providing a ProposalStore upon request. The trait extends Clone, Sync
and Send so it can easily be moved between threads and into new
contexts.

This commit also updates the code usages of ProposalStore to accept a
ProposalStoreFactory and use that to get the ProposalStore.

Signed-off-by: Caleb Hill <hill@bitwise.io>