Make proposal_store module public
The usages of `ProposalStore` and `ProposalStoreFactory` used to be
contained to libsplinter but now that the admin endpoints are moving to
the rest_api/actix_web_1 crate those types are used there as well. This
commit changes the visibility of the `proposal_store` module so it can
be used from rest_api/actix_web_1.

Signed-off-by: Caleb Hill <hill@bitwise.io>