Update Dockerfiles to copy more than Cargo.tomls
The Dockerfile-installed-focal,
examples/gameroom/daemon/Dockerfile-installed-focal and tests/Dockerfile
copy just the Cargo.tomls for the rest_api crates. Which is fine as long
as no other crates depend on those rest_api crates. As soon as splinterd
pulls in the actix_web_1 or 4 crates cargo won't be able to find those
local dependencies.

This commit updates the dockerfiles to copy the entire actix_web_1,
actix_web_4, and common rest_api crates into the build directory.

Signed-off-by: Caleb Hill <hill@bitwise.io>