Add NotifyObserver trait and command implementation
Notifications from consensus will need to be handled by
different components of the service. The NotifyObserver
trait provides the API for handling these notifications.

The current implementation of an observer is for the
2pc prototype. The notifications result in an CommitEntry
being stored in state, getting set to either committed or
aborted. On RequestForStart, a random value will be generated
and used for agreement.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>