Convert From to TryFrom for CircuitStatus
A `From` implementation does not accurately convert between string
representations of `CircuitStatus`s and actual `CircuitStatus`s. All
strings that are not `disbanded` or `abandoned` do not logically
map to `CircuitStatus::Active`. This commit changes the `From`
implementation to a `TryFrom` implementation and updates the single use
of the `from` function.

Signed-off-by: Caleb Hill <hill@bitwise.io>