Update scabbard store consensus context
Update the `Context` struct in scabbard_store to flatten the
`role_context` field. This commit removes the `role_context` field,
which was specific to a participant or coordinator, and adds two fields:
`participants` and `state`. Previously participants and state were
part of the role context.

This commit also removes the `CoordinatorContext` and `ParticipantContext`
structs that were used in role context. The added `state` field in
`Context` has type `Scabbard2pcState` which is a combination of all
variants in the `CoordinatorContext` and `ParticiantContext` structs.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>