Update scabbard store context and action migrations
Update diesel migrations for the context and action conensus tables
to combine participant and coordinator.

Previously participant and coordinator consensus context and actions
were separated and stored in different tables because the associated
data differed slightly between the two. This commit combines the tables
in the diesel migrations to create generic context and action tables
that can hold all necessary information for both coordinators and
participants.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>