Combine participant and coordinator tables
Combine the separate coordinator and participant tables for consensus
action and context in the scabbard store.

This commit updates the schema, models, and operations for the diesel
backed implementation of `ScabbardStore` to combine the coordinator
and participant tables into generic tables that can hold the
necessary data for both types of members.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>