Update scabbard_store_add_event test
Update the the last action in the scabbard_store_add_event test
which attempted to get an error trying to add a coordinator
specific event for a participant context. This will no longer
produce an error because the context tables have been combined,
so the add_consensus_event method is not aware of whether the
context belongs to a participant or coordinator and therefore
will add any valid event.

Update the last action in the test to attempt to add an event
for a service that does not exist and check that an error is
returned.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>