Add set_alarm operation to scabbard store
Add a `set_alarm` method to the `ScabbardStore` trait and
implementations. This method sets an alarm of a specified type for the
given service. If there is an existing alarm of the same type already
set for the given service the existing alarm will be overwritten with
the new alarm.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>