Change From impl to new constructor
The `From` trait documentation states that the trait is "Used to do
value-to-value conversions while consuming the input value." the
implementation this commit replaces did not consume the `&AdminService`.

This commit replaces the `from` method with a `pub new` method that does
exactly the same thing while more strictly adhering to the spirit of
the `From` trait.

Signed-off-by: Caleb Hill <hill@bitwise.io>