Update `list_ready_services` to check events
Update the `list_ready_services` operation in the diesel scabbard store
implementation to check the events table and list any services which
have outstanding events.

Update the `test_scabbard_timer_filter` test to add an unexecuted event
for one of the services and check that a service is listed as "ready"
because of it.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>