Add AlarmType enum
Add an `AlarmType` enum in an `Alarm` module. This enum currently has
one variant: TwoPhaseCommit, this is the alarm type for two phase
commit consensus alarms. When alarms are added to the database
for a service an alarm type is required, this alarm type allows
components to differentiate between the alarms set for a specific
service.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>